+++ pyespn/classes/__init__.py
@@ -1,6 +1,8 @@
 from .team import Team
-from .player import Player
+from .player import Player, Recruit
 from .event import Event
 from .venue import Venue
 from .league import League
 from .schedule import Schedule
+from .draft import DraftPick
+from .vehicle import Vehicle