+++ pyespn/classes/player.py
@@ -1,4 +1,5 @@
 from pyespn.core.decorators import validate_json
+from pyespn.classes.vehicle import Vehicle
 
 
 @validate_json('player_json')
@@ -15,8 +16,9 @@ class Player:
         api_ref (str | None): API reference link for the player.
         id (str | None): The unique identifier for the player.
         uid (str | None): The ESPN UID for the player.
+        flag (dict | None): a dict with players nationality/flag (mostly for racing).
         guid (str | None): The GUID associated with the player.
-        type (str | None): The type of player (e.g., "athlete").
+        type (str | None): The type of player (e.g., 'athlete').
         alternate_ids (str | None): Alternative ID for the player.
         first_name (str | None): The player's first name.
         last_name (str | None): The player's last name.
@@ -77,9 +79,9 @@ class Player:
         Returns a string representation of the Player instance.
 
         Returns:
-            str: A formatted string with the players's name, debut year and jersey.
+            str: A formatted string with the players's name, position and jersey.
         """
-        return f"<Player | {self.full_name}, {self.debut_year} ({self.jersey})>"
+        return f"<Player | {self.full_name}, {self.position_abbreviation} ({self.jersey})>"
 
     def _set_player_data(self):
         """
@@ -90,6 +92,7 @@ class Player:
         self.uid = self.player_json.get('uid')
         self.guid = self.player_json.get('guid')
         self.type = self.player_json.get('type')
+        self.flag = self.player_json.get('flag')
         self.alternate_ids = self.player_json.get('alternateIds', {}).get('sdr')
         self.first_name = self.player_json.get('firstName')
         self.last_name = self.player_json.get('lastName')
@@ -103,6 +106,7 @@ class Player:
         self.age = self.player_json.get('age')
         self.date_of_birth = self.player_json.get('dateOfBirth')
         self.debut_year = self.player_json.get('debutYear')
+        self.college_athlete_ref = self.player_json.get('collegeAthlete', {}).get('$ref')
 
         self.links = self.player_json.get('links', [])
 
@@ -112,7 +116,7 @@ class Player:
 
         self.college_ref = self.player_json.get('college', {}).get('$ref')
         self.slug = self.player_json.get('slug')
-        self.jersey = self.player_json.get('jersey')
+        self.jersey = self.player_json.get('jersey', '--')
 
         position = self.player_json.get('position', {})
         self.position_ref = position.get('$ref')
@@ -140,6 +144,12 @@ class Player:
         self.status_abbreviation = status.get('abbreviation')
 
         self.statistics_log_ref = self.player_json.get('statisticslog', {}).get('$ref')
+
+        if 'vehicles' in self.player_json:
+            self.vehicles = []
+            for vehicle in self.player_json.get('vehicles'):
+                self.vehicles.append(Vehicle(vehicle_json=vehicle,
+                                             espn_instance=self.espn_instance))
 
     def to_dict(self) -> dict:
         """
@@ -255,7 +265,6 @@ class Recruit:
         self.high_school_id = high_school.get('id')
         self.high_school_name = high_school.get('name')
         self.high_school_state = high_school.get('address', {}).get('state')
-
 
         self.slug = self.recruit_json.get('slug')