+++ pyespn/core/client.py
@@ -5,7 +5,11 @@ from pyespn.data.betting import (BETTING_PROVIDERS, DEFAULT_BETTING_PROVIDERS_MA
                                  LEAGUE_DIVISION_FUTURES_MAPPING)
 from pyespn.exceptions import API400Error
 from .decorators import *
+from typing import TYPE_CHECKING
 import concurrent.futures
+
+if TYPE_CHECKING:
+    from pyespn.classes import Team, Player, Recruit, Event, League  # Only imports for type checking
 
 
 @validate_league
@@ -56,6 +60,7 @@ class PYESPN:
         self.schedules = {}
         self.recruit_rankings = {}
         self.drafts = {}
+        self.athletes = {}
         self.league = None
         self._load_league_data()
         if load_teams:
@@ -136,12 +141,25 @@ class PYESPN:
 
         self.schedules[season] = self.get_regular_seasons_schedule(season=season)
 
-    def load_year_draft(self, season: int):
+    def load_year_draft(self, season: int) -> None:
+        """
+        Loads draft data for a given season and stores it in the drafts dictionary.
+
+        This method retrieves draft data for the specified season using
+        `load_draft_data_core` and associates it with the season in the `drafts` attribute.
+
+        Args:
+            season (int): The season year for which to load draft data.
+
+        Returns:
+            None
+        """
+
         self.drafts[season] = load_draft_data_core(season=season,
                                                    league_abbv=self.league_abbv,
                                                    espn_instance=self)
 
-    def get_player_info(self, player_id) -> dict:
+    def get_player_info(self, player_id) -> "Player":
         """
         Retrieves detailed information about a player.
 
@@ -149,7 +167,7 @@ class PYESPN:
             player_id (str): The ID of the player.
 
         Returns:
-            dict: The player's information.
+            Player: The player's information in player class
         """
         return get_player_info_core(player_id=player_id,
                                     league_abbv=self.league_abbv,
@@ -165,7 +183,7 @@ class PYESPN:
         return get_player_ids_core(league_abbv=self.league_abbv)
 
     @requires_college_league('recruiting')
-    def get_recruiting_rankings(self, season, max_pages=None) -> dict:
+    def get_recruiting_rankings(self, season, max_pages=None) -> list["Recruit"]:
         """
         Retrieves the recruiting rankings for a given season.
 
@@ -174,7 +192,7 @@ class PYESPN:
             max_pages (int, optional): The maximum number of pages of data to retrieve.
 
         Returns:
-            dict: The recruiting rankings.
+            list[Recruit]: The recruiting rankings.
         """
         return get_recruiting_rankings_core(season=season,
                                             league_abbv=self.league_abbv,
@@ -191,7 +209,7 @@ class PYESPN:
 
         self.recruit_rankings = {year: self.get_recruiting_rankings(season=year)}
 
-    def get_game_info(self, event_id) -> dict:
+    def get_game_info(self, event_id) -> "Event":
         """
         Retrieves detailed information about a specific game.
 
@@ -199,13 +217,13 @@ class PYESPN:
             event_id (str): The ID of the game.
 
         Returns:
-            dict: The game's information.
+            Event: The game's information.
         """
         return get_game_info_core(event_id=event_id,
                                   league_abbv=self.league_abbv,
                                   espn_instnace=self)
 
-    def get_team_info(self, team_id) -> dict:
+    def get_team_info(self, team_id) -> "Team":
         """
         Retrieves detailed information about a team.
 
@@ -213,7 +231,7 @@ class PYESPN:
             team_id (str): The ID of the team.
 
         Returns:
-            dict: The team's information.
+            Team: The team's information.
         """
         return get_team_info_core(team_id=team_id,
                                   league_abbv=self.league_abbv,
@@ -471,12 +489,12 @@ class PYESPN:
                                   standings_type=standings_type,
                                   league_abbv=self.league_abbv)
 
-    def get_league_info(self) -> dict:
+    def get_league_info(self) -> "League":
         """
         Retrieves information about the league.
 
         Returns:
-            dict: The league's information.
+            League: The league's information.
         """
         return get_league_info_core(league_abbv=self.league_abbv,
                                     espn_instance=self)
@@ -495,7 +513,7 @@ class PYESPN:
                                                 espn_instance=self,
                                                 season=season)
 
-    def get_team_by_id(self, team_id):
+    def get_team_by_id(self, team_id) -> "Team":
         """
         Finds and returns the Team object that matches the given team_id.
 
@@ -506,3 +524,49 @@ class PYESPN:
             Team: The matching Team object, or None if not found.
         """
         return next((team for team in self.teams if str(team.team_id) == str(team_id)), None)
+
+    def load_season_rosters(self, season):
+        """
+        Loads the season roster for all teams in the league.
+
+        This method iterates through all teams and calls their `load_season_roster`
+        method to fetch and store the roster data for the specified season.
+
+        Args:
+            season (int or str): The season year for which to load rosters.
+
+        Returns:
+            None
+
+        Example:
+            >>> espn = PYESPN('nfl')
+            >>> espn.load_season_rosters(season=2023)
+            >>> for team in espn.teams:
+            >>>     print(team.roster[2023])
+            [<Player | John Doe>, <Player | Jane Smith>, ...]
+
+        """
+
+        for team in self.teams:
+            team.load_season_roster(season=season)
+
+    def load_athletes(self, season) -> None:
+        """
+        Loads and stores athlete data for a given season.
+
+        This function retrieves athlete data for the specified season using `load_athletes_core`
+        and stores it in the `athletes` attribute of the instance.
+
+        Args:
+            season (int): The season year for which athlete data is being loaded.
+
+        Returns:
+            None: The retrieved athlete data is stored in `self.athletes[season]`.
+
+        Notes:
+            - Uses `load_athletes_core` to fetch athlete data.
+            - Stores the result in `self.athletes` with the season as the key.
+        """
+        self.athletes[season] = load_athletes_core(season=season,
+                                                   league_abbv=self.league_abbv,
+                                                   espn_instance=self)